// src/app/core/models/members/profile-education.model.ts
export class MemberProfileEducation {
  schoolID: string = '';
  schoolName: string = '';
  schoolImage: string = '';
  schoolAddress: string = '';
  major: string = '';
  degree: string = '';
  degreeTypeID: string = '';
  yearClass: string = '';
  schoolType: string = '';
  Societies: string = '';
  webSite: string = '';
  sportLevelType: string = '';
}